Simplify extensible services' usage of mixins
